This PR improves the overall query performance for all databases.

The improvements come from two changes to the Oplog.

1) Remove an unnecessary entry encoding step

In Entry's `decode()` function we had it so that we encoded the entry once the serialized block was decoded. This was unnecessary and by removing this step we gain a significant write and read performance improvement. From:
```
Starting benchmark...
Insert 1000 documents
Inserting 1000 documents took 1642 ms, 609 ops/s, 1.642 ms/op
Query 1000 documents
Querying 1000 documents took 72 ms, 13888 ops/s, 0.072 ms/op
```
To:
```
Starting benchmark...
Insert 1000 documents
Inserting 1000 documents took 1284 ms, 778 ops/s, 1.284 ms/op
Query 1000 documents
Querying 1000 documents took 52 ms, 19230 ops/s, 0.052 ms/op
```

2) Remove `useRefs` in oplog's traverse() function

We don't need to use the `refs` anymore in `traverse` as we changed the entry loading logic a while back. Removing it improves the query performance time from:
```
Querying 1000 documents took 52 ms, 19230 ops/s, 0.052 ms/op
```
To:
```
Querying 1000 documents took 27 ms, 37037 ops/s, 0.027 ms/op
```

Altogether with these two changes, the performance improvement is:

Before:
```
Inserting 1000 documents took 1642 ms, 609 ops/s, 1.642 ms/op
Querying 1000 documents took 72 ms, 13888 ops/s, 0.072 ms/op
```

After:
```
Inserting 1000 documents took 1284 ms, 778 ops/s, 1.284 ms/op
Querying 1000 documents took 27 ms, 37037 ops/s, 0.027 ms/op
```

This PR is on top of #1176.